Fix issue of messages not being added to the top when deleting
Deleting messages wouldn't add messages to the top in response to the
number of rendered messages falling below the renderedMax value.

Probably not that big a deal, as they would still get added when scrolling
up.  Except if enough messages were deleted to make the scroll bar go away,
in which case there was no way to scroll up!