Fix problem where scrolling up quickly stops the lazy loading from working
This happens when scrolling up quickly vs how long it takes to render chat
messages.  When the new messages are rendered, instead of keeping the same
scroll position, it instantly scrolls to the top of the newly added messages.
Then it won't render more messages until the window is scrolled down some first.

This is a foundry core issue https://github.com/foundryvtt/foundryvtt/issues/11129

The problem is that the scroll position isn't maintained if it's at 0 exactly
when the new messages are added.  Just set it to 1 in that case, then it will
scroll.