Add publish to foundry workflow
This runs when the release is changed from a draft release to a published
release.  It will add it the foundry page for the module.

Not running immediatly gives a chance to verify the module.zip and json files
are correct before making the release live.  And the release message on github
can be edited.