Fix issue with duplicate messages appearing
The root cause was that deleting messages would do a 100 ms slide up animation,
then delete the message html from the chatlog when the animation finished.
This was done asynchronously.  Until the animation finished, the message was
still in the chat log html list.

While the animation was happening, another operation (such as deleting a
different message) might want to update the variable holding the message ID of
the last rendered message in the chatlog.  It would find the message that was in
the process of being deleted and set it as the last message in the log.

The last message would then be set to a message that didn't exist, which caused
all sort of problems, often leading to duplicate messages.